requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
openai==1.3.8
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
stripe==7.8.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
slowapi==0.1.9
email-validator==2.1.0
